useless warnings about call-parameter type mismatches

+++ vu/vmadh.h
@@ -1,12 +1,12 @@
 #include "vu.h"
 
-INLINE void do_madh(short* VD, signed short* VS, signed short* VT)
+INLINE void do_madh(short* VD, short* VS, short* VT)
 {
     INT64 acc[N];
     register int i;
 
     for (i = 0; i < N; i++)
-        acc[i] = VS[i] * VT[i];
+        acc[i] = (signed short)(VS[i]) * (signed short)(VT[i]);
     for (i = 0; i < N; i++)
         acc[i] = acc[i] << 16;
     do_acc(acc);
